reword some sentences about scripting
My brain is starting to fuzz out.  It's creeping up on 1am.  I can
already tell dragging out of bed in the morning is going to be pure hell.

I wonder if it's going to be snowing.  I hope not.  I want to ride my
bike to work.

It will probably be snowing.  Or raining.  Or there will be a plague of
frogs or something.